+++ src/extensions/rules/rulesext.cpp
@@ -114,7 +114,8 @@ RulesClassExtension::RulesClassExtension(const RulesClass* this_ptr) :
     IronCurtainRechargeTime(9900),
     IronCurtainFlashRate(8),
     IronCurtainFlashIntensityMultiplier(50),
-    IronCurtainSound(VOC_NONE)
+    IronCurtainSound(VOC_NONE),
+    ComesNearWaypointDistance(CELL_LEPTON_W * 5)
 {
     //if (this_ptr) EXT_DEBUG_TRACE("RulesClassExtension::RulesClassExtension - 0x%08X\n", (uintptr_t)(ThisPtr));
 
@@ -276,6 +277,7 @@ void RulesClassExtension::Object_CRC(CRCEngine &crc) const
     crc(IronCurtainDuration);
     crc(IronCurtainRechargeTime);
     crc(IronCurtains.Count());
+    crc(ComesNearWaypointDistance);
 }
 
 
@@ -719,6 +721,7 @@ bool RulesClassExtension::General(CCINIClass &ini)
     }
 
     IsBeachIsCrush = ini.Get_Bool(GENERAL, "BeachIsCrush", IsBeachIsCrush);
+    ComesNearWaypointDistance = ini.Get_Int(GENERAL, "ComesNearWaypointDistance", ComesNearWaypointDistance);
 
     IronCurtains = ::TGet_TypeList(ini, GENERAL, "IronCurtains", IronCurtains);
     IronCurtainDuration = ini.Get_Int(GENERAL, "IronCurtainDuration", IronCurtainDuration);