+++ docs/Whats-New.md
@@ -86,6 +86,7 @@ New:
 - Add "Create Building At" trigger action (by Rampastring)
 - Add the Iron Curtain logic from Red Alert 1 for map scripting and AI (by Rampastring)
 - Heap dumping command now logs more information of Techno objects (by Rampastring)
+- Allow customizing the distance for the "Comes Near Waypoint" trigger event (by Rampastring)
 
 Vinifera fixes:
 - Fix unit placement in non-TS Client builds of Vinifera (by ZivDero)